requirements.txt
torch
transformers
Pillow
opencv-python-headless
numpy
tqdm
torchvision
rembg
onnxruntime
easydict